+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -24,7 +24,7 @@ public interface StudentService {
 
     Integer getAmountOfStudents ();
 
-    int getAverageAge ();
+    Double getAverageAge ();
     List<Student> getLastStudents();
 
 }
